Massive overhaul and PEP8 compatibility fixes
