+++ modules/seminar/format/format-round.tex
@@ -3,12 +3,12 @@
 
 \fancypagestyle{seminar-(* competition.id *)-(* volume.id *)-(* semester.id *)-(* round.id *)-problems}{%
     \pagestyle{seminar-main-(* competition.id *)}
-    \fancyhead[LE,RO]{\textit{Zadania (* round.number *). kola (* semester.genitive *) časti -- (* round.deadline.strftime('%d. %m. %Y') *)}}
+    \fancyhead[LE,RO]{\textit{Zadania (* round.number *). kola (* semester.feminine.genitive *) časti -- (* round.deadline.strftime('%d. %m. %Y') *)}}
     \titleformat{\subsection}[hang]{\large}{\bfseries\arabic{round}.\arabic{problem}}{6pt}{}[]
 }
 
 \fancypagestyle{seminar-(* competition.id *)-(* volume.id *)-(* semester.id *)-(* round.id *)-solutions}{%
     \pagestyle{seminar-main-(* competition.id *)}
-    \fancyhead[LE,RO]{\textit{Riešenia (* round.number *). kola (* semester.genitive *) časti}}
+    \fancyhead[LE,RO]{\textit{Riešenia (* round.number *). kola (* semester.feminine.genitive *) časti}}
     \titleformat{\subsection}[hang]{\large}{\bfseries\arabic{round}.\arabic{problem}}{6pt}{}[]
 }
