+++ modules/seminar/module.mk
@@ -72,13 +72,13 @@ input/seminar/%/problems.tex input/seminar/%/solutions.tex: \
 #	@mkdir -p $(dir $@)
 #	python3 ./modules/seminar/build-semester.py 'source/seminar/' 'modules/seminar' $(word 1,$(words)) $(word 2,$(words)) $(word 3,$(words)) -o '$(dir $@)'
 #
-#input/seminar/%/invite.tex: \
-#	modules/seminar/templates/invite.tex \
-#	input/seminar/$$*/format/semester.tex \
-#	source/seminar/$$*/meta.yaml
-#	$(eval words := $(subst /, ,$*))
-#	@mkdir -p $(dir $@)
-#	python3 modules/seminar/build-invite.py 'source/seminar/' 'modules/seminar' $(word 1,$(words)) $(word 2,$(words)) $(word 3,$(words)) -o '$(dir $@)'
+input/seminar/%/invite.tex: \
+	modules/seminar/templates/invite.tex \
+	input/seminar/$$*/format-semester.tex \
+	source/seminar/$$*/meta.yaml
+	$(eval words := $(subst /, ,$*))
+	@mkdir -p $(dir $@)
+	python3 modules/seminar/build/invite.py 'source/seminar/' 'modules/seminar' -c $(word 1,$(words)) -v $(word 2,$(words)) -s $(word 3,$(words)) -o '$(dir $@)'
 
 input/seminar/%/pdf-prerequisites: \
 	$$(subst source/,input/,$$(wildcard source/seminar/$$*/*/*.jpg)) \
@@ -96,59 +96,59 @@ output/seminar/%/html-prerequisites: \
 
 output/seminar/%/problems.pdf: \
 	$$(subst source/,input/,$$(subst .md,.tex,$$(wildcard source/seminar/$$*/*/problem.md))) \
-	input/seminar/%/pdf-prerequisites \
-	input/seminar/%/problems.tex
+	input/seminar/$$*/pdf-prerequisites \
+	input/seminar/$$*/problems.tex
 	$(call doubletex,seminar)
 
 output/seminar/%/solutions.pdf: \
 	$$(subst source/,input/,$$(subst .md,.tex,$$(wildcard source/seminar/$$*/*/solution.md))) \
-	input/seminar/%/pdf-prerequisites \
-	input/seminar/%/solutions.tex
+	input/seminar/$$*/pdf-prerequisites \
+	input/seminar/$$*/solutions.tex
 	$(call doubletex,seminar)
 
 output/seminar/%/semester.pdf: \
 	$$(subst source/,input/,$$(subst .md,.tex,$$(wildcard source/seminar/$$*/*/*/problem.md))) \
-	input/seminar/%/*/pdf-prerequisites \
-	input/seminar/%/semester.tex
+	input/seminar/$$*/*/pdf-prerequisites \
+	input/seminar/$$*/semester.tex
+	$(call doubletex,seminar)
+
+output/seminar/%/invite.pdf:\
+	source/seminar/$$*/meta.yaml \
+	input/seminar/$$*/invite.tex
 	$(call doubletex,seminar)
 
 output/seminar/%/semester-print.pdf: \
-	output/seminar/%/semester.pdf
+	output/seminar/$$*/semester.pdf
 	@echo -e '$(c_action)Converting $(c_filename)$<$(c_action) to a short-edge booklet $(c_filename)$@$(c_action):$(c_default)'
 	pdfbook --short-edge --quiet --outfile $@ $<
 
 output/seminar/%/html-problems: \
-	output/seminar/%/html-prerequisites \
+	output/seminar/$$*/html-prerequisites \
 	$$(subst source/,output/,$$(subst .md,.html,$$(wildcard source/seminar/$$*/*/problem.md))) ;
 	
 output/seminar/%/html-solutions:\
-	output/seminar/%/html-prerequisites \
+	output/seminar/$$*/html-prerequisites \
 	$$(subst source/,output/,$$(subst .md,.html,$$(wildcard source/seminar/$$*/*/solution.md))) ;
 
 output/seminar/%/pdf: \
-	output/seminar/%/problems.pdf \
-	output/seminar/%/solutions.pdf ;
+	output/seminar/$$*/problems.pdf \
+	output/seminar/$$*/solutions.pdf ;
 
 output/seminar/%/html: \
-	output/seminar/%/html-problems \
-	output/seminar/%/html-solutions ;
+	output/seminar/$$*/html-problems \
+	output/seminar/$$*/html-solutions ;
 
 output/seminar/%/problems: \
-	output/seminar/%/problems.pdf \
-	output/seminar/%/html-problems ;
+	output/seminar/$$*/problems.pdf \
+	output/seminar/$$*/html-problems ;
 
 output/seminar/%/solutions: \
-	output/seminar/%/solutions.pdf \
-	output/seminar/%/html-solutions ;
+	output/seminar/$$*/solutions.pdf \
+	output/seminar/$$*/html-solutions ;
 
 output/seminar/%: \
-	output/seminar/%/problems \
-	output/seminar/%/solutions ;
-
-output/seminar/%/invite.pdf:\
-	input/$$*/invite.tex\
-	source/%/meta.yaml
-	$(call doubletex,seminar)
+	output/seminar/$$*/problems \
+	output/seminar/$$*/solutions ;
 
 .PHONY:
