+++ modules/seminar/templates/problems.tex
@@ -5,7 +5,7 @@
 (@ endblock pagestyle @)
 
 (@ block inner @)
-    \section{\texorpdfstring{Zadania (* round.id *). kola (* semester.genitive *) časti}{Zadania}}
+    \section{\texorpdfstring{Zadania (* round.id *). kola (* semester.feminine.genitive *) časti}{Zadania}}
     {
         \centering
         \vspace*{-5mm}
@@ -24,7 +24,7 @@
             \texorpdfstring{%
                 \large \textbf{(* problem.title *)}%
                 (@ if problem.categories @)%
-                \normalsize \hfill kategóri(@ if problem.categories|length == 1 @)a(@ else @)e(@ endif @) (* problem.categories|formatList *)%
+                \normalsize \hfill kategóri(@ if problem.categories|length == 1 @)a(@ else @)e(@ endif @) (* problem.categories|format_list *)%
                 (@ endif @)
             }{%
                 (* round.number *).(* problem.number *) (* problem.title *)%
